Optimize code
Make not append 1 char in each loop if not comment or comment end but
save the start index as it was in old code without this PR and then copy
to buffer only if necessary. Also restore old behavior of not ingnoring
the complete statement if it has unclosed quote at file end, so that as
before with this PR people will see the sql error and not see nothing
like with this PR before this commit.